separating ViewPosts logic and UI
